+++ zb_lrcon.c
@@ -79,7 +79,7 @@ qboolean ReadLRconFile(char *lrcname)
 			int len;
 			
 			// remove '\n'
-			len = q2a_strlen(buffer) - 1;
+			len = (int)q2a_strlen(buffer) - 1;
 			if(buffer[len] == '\n')
 				{
 					buffer[len] = 0x0;
@@ -141,7 +141,7 @@ qboolean ReadLRconFile(char *lrcname)
 					SKIPBLANK(cp);
 					*pp = 0;
 					
-					len = q2a_strlen(cp) + 1;
+					len = (int)q2a_strlen(cp) + 1;
 					
 					// zero length command
 					if(!len)
@@ -460,7 +460,7 @@ void lrconRun(int startarg, edict_t *ent, int client)
 			return;
 		}
 		
-	len = q2a_strlen(cmd) + 1;
+	len = (int)q2a_strlen(cmd) + 1;
 	
 	lrconcmds[maxlrcon_cmds].password = gi.TagMalloc (len, TAG_LEVEL);
 	
@@ -476,7 +476,7 @@ void lrconRun(int startarg, edict_t *ent, int client)
 			return;
 		}
 		
-	len = q2a_strlen(cmd) + 20;
+	len = (int)q2a_strlen(cmd) + 20;
 	
 	lrconcmds[maxlrcon_cmds].lrconcmd = gi.TagMalloc (len, TAG_LEVEL);
 	processstring(lrconcmds[maxlrcon_cmds].lrconcmd, cmd, len - 1, 0);
